Remove S4 bc it wasn't asynchronous anyway
