+++ src/app/upload.rs
@@ -68,7 +68,7 @@ pub fn handle_multipart_item(
 pub fn upload(req: HttpRequest<State>) -> FutureResponse<HttpResponse> {
     use actix::Addr;
     use actix_web::error;
-    use crate::object_store::ObjectStore;
+    use crate::object::ObjectStore;
     let store: Addr<ObjectStore> = req.state().store.clone();
 
     use crate::sessions::UserSession;
