+++ src/object/mod.rs
@@ -1,3 +1,5 @@
+pub mod store;
+pub use store::ObjectStore;
 
 mod object_id;
 pub use object_id::ObjectId;
@@ -5,6 +7,5 @@ pub use object_id::ObjectId;
 mod object_row;
 pub use object_row::ObjectRow;
 
-pub trait Object {
-    fn id(&self) -> ObjectId;
-}
+mod object;
+pub use object::Object;
