+++ src/object/object.rs
@@ -1,0 +1,5 @@
+use super::ObjectId;
+
+pub trait Object {
+    fn id(&self) -> ObjectId;
+}
