--- src/store/mod.rs
@@ -1,2 +1,0 @@
-mod s4;
-pub mod object_store;
