--- src/store/s4/README.md
@@ -1,9 +1,0 @@
-# S4
-
-This is forked from another crate, called `s4`.
-
-That crate does not compile, and looks to be unmaintained,
-but the trait and implementations greatly simplify the interactions
-with Rusoto S3, and give us a strong headstart on interacting with S3.
-
-We started from where they left off.
