+++ templates/upload.html.j2
@@ -9,19 +9,31 @@
 <h1>Collect</h1>
 <form enctype="multipart/form-data" method="POST">
     <input type="file" name="file">
-    <input type="button" onclick="upload(this)">
+    <button type="button" onclick="upload(this)">Upload</button>
+    <div class="progress"></div>
 </form>
+<br>
+<a href="/static/uploads">Uploads</a>
 <script>
   function upload(input){
     var xhr = new XMLHttpRequest();
+    var formElt = input.parentElement;
+    var progressElt = formElt.querySelector(".progress");
     xhr.upload.onprogress = function(e) {
-      console.log(e.loaded, e.total)
+      // progress uploading
+      let amountDone = e.loaded / e.total;
+      let percentage = Math.round(1000 * amountDone) * 0.1;
+      progressElt.innerText = String(percentage).slice(0,4) + "%";
     }
     xhr.upload.onload = function(e) {
-      console.log('file upload')
+      // is done uploading
+      console.log('file uploaded')
+      progressElt.innerText = "100% Complete";
+      formElt.file.value = null;
     }
     xhr.open("POST", "/upload", true);
-    xhr.send(new FormData(input.parentElement));
+    xhr.send(new FormData(formElt));
+    return false;
   }
 </script>
 {% endblock %}
